requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
eventlet==0.30.2
pyzmq==22.3.0
python-engineio==4.3.0
dnspython==1.16.0